+++ src/ServiceContracts/PayoutsContract.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Dodopayments\ServiceContracts;
+
+use Dodopayments\DefaultPageNumberPagination;
+use Dodopayments\Payouts\PayoutListResponse;
+use Dodopayments\RequestOptions;
+
+use const Dodopayments\Core\OMIT as omit;
+
+interface PayoutsContract
+{
+    /**
+     * @api
+     *
+     * @param int $pageNumber Page number default is 0
+     * @param int $pageSize Page size default is 10 max is 100
+     *
+     * @return DefaultPageNumberPagination<PayoutListResponse>
+     */
+    public function list(
+        $pageNumber = omit,
+        $pageSize = omit,
+        ?RequestOptions $requestOptions = null,
+    ): DefaultPageNumberPagination;
+}